feat: Integrate commit details into a new info pane within the commit graph view, replacing the separate changed files view.
