+++ src/views/CommitGraphViewProvider.ts
@@ -4,7 +4,7 @@
 
 import * as vscode from "vscode";
 import { GitOps } from "../git/operations";
-import type { Branch } from "../types";
+import type { Branch, CommitDetail } from "../types";
 import type {
     BranchAction,
     CommitAction,
@@ -24,6 +24,7 @@ export class CommitGraphViewProvider implements vscode.WebviewViewProvider {
     private readonly PAGE_SIZE = 500;
 
     private branches: Branch[] = [];
+    private selectedCommitDetail: CommitDetail | null = null;
 
     private readonly _onCommitSelected = new vscode.EventEmitter<string>();
     readonly onCommitSelected = this._onCommitSelected.event;
@@ -68,6 +69,7 @@ export class CommitGraphViewProvider implements vscode.WebviewViewProvider {
                 case "ready":
                     this.sendBranches();
                     await this.loadInitial();
+                    this.postCommitDetailState();
                     break;
                 case "selectCommit":
                     this._onCommitSelected.fire(msg.hash);
@@ -117,6 +119,16 @@ export class CommitGraphViewProvider implements vscode.WebviewViewProvider {
     async refresh(): Promise<void> {
         this.sendBranches();
         await this.loadInitial();
+    }
+
+    setCommitDetail(detail: CommitDetail): void {
+        this.selectedCommitDetail = detail;
+        this.postCommitDetailState();
+    }
+
+    clearCommitDetail(): void {
+        this.selectedCommitDetail = null;
+        this.postCommitDetailState();
     }
 
     private sendBranches(): void {
@@ -185,6 +197,14 @@ export class CommitGraphViewProvider implements vscode.WebviewViewProvider {
 
     private postToWebview(msg: CommitGraphInbound): void {
         this.view?.webview.postMessage(msg);
+    }
+
+    private postCommitDetailState(): void {
+        if (this.selectedCommitDetail) {
+            this.postToWebview({ type: "setCommitDetail", detail: this.selectedCommitDetail });
+            return;
+        }
+        this.postToWebview({ type: "clearCommitDetail" });
     }
 
     private getHtml(webview: vscode.Webview): string {
